Remove unused localleader from mappings.lua
For the use case of including the Nvim-R plugin for working with the R
programming language the plugin makes extensive use of the LocalLeader
mappings. As both the LocalLeader and Global Leader keys are mapped to
" " in Astrovim this introduces conflicts with the existing keybindings when
the filetype plugin is active. As there appears to be no reference to
the localleader mapping throughout the Astrovim repo by removing the
definition the plugin works properly and Astrovim reports no errors
after the change.