Fix error when trying to access properties of non-valid buffer in is_valid buffer function
If you try to access `vim.bo[bufnr].buflisted` of an invalid `bufnr` the editor will throw an error. This happens when special buffers are defined but are not valid.

Putting `vim.api.nvim_buf_is_valid(bufnr)` first in the condition makes sure the call to `vim.bo` doesn't throw an error.

# How to reproduce the error?

* Open nvim (without parameters) and load any file (type is irrelevant)
* Close nvim
* Open nvim  (without parameters) and use `Ctrl-o` twice. 

## Why `Ctrl-o` twice?

This is an easy way to jump back directly to the last file you were modifying.