fix(cmp): only build LuaSnip if NOT on Windows
`make` is usually available on UNIX systems, but the current check sets the `build` property if `win32` IS PRESENT.
This means that the `build` property is present in the case of Windows, and absent in the case of UNIX.
As a Windows user, I get an error every time I update my plugins, and UNIX users don't get the feature they usually can install correctly.
I updated the check to detect if `win32` is NOT `1`, instead of NOT `0`, so that the `build` property is only present on non-Windows systems.