fix: closing last buffer in a tabpage closes the tab
Previously, closing a the last buffer resulted in the next loaded buffer
in the jumplist being displayed in the current tabpage window.
The fix checks if there are less than 2 pages in the tab. If yes, save
the current buffer_number, close the tab, and proceed to close the
buffer(which remains hidden at this point). The next tab's window is
selected, depending on the jumplist.