feat(lsp): symbol search for word under cursor (#2199)
* feat(lsp): queried word in symbol search's title

While performing a workspace symbols search (telescope's
builtin.lsp_workspace_symbols()), the user's inputed query is now shown as
a prompt title in the telescope's window.

* feat(lsp): symbol search for word under cursor

If the given word for a workspace symbol search (telescope's
builtin.lsp_workspace_symbols()) is empty, takes the word under the
cursor as the query parameter.